requirements.txt
requests
numpy
pillow
pathos
